requirements.txt
datetime==4.3.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-RESTful==0.3.9
flask-restplus==0.13.0
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
gunicorn==20.0.4
Jinja2==2.11.3
#loguru==0.5.3
MarkupSafe==1.1.1
passlib==1.7.4
python-dotenv==0.15.0
requests==2.25.1
SQLAlchemy==1.3.18
urllib3==1.26.5
Werkzeug==1.0.1
WTForms==2.3.3
# XlsxWriter==1.4.2
